move MockTick
