+++ staging_vespalib/src/tests/metrics/mock_tick.cpp
